Moved functions to remove circular dependency.
